Update chat/contents api endpoints
