+++ backend/src/app.ts
@@ -11,7 +11,6 @@ import apiKeys from "./routes/apiKeys";
 import gpi from "./routes/gpi";
 import gpis from "./routes/gpis";
 import chat from "./routes/chat";
-import chatContent from "./routes/chatContent";
 import chatContents from "./routes/chatContents";
 import histories from "./routes/histories";
 import providerTypes from "./routes/providerTypes";
@@ -46,7 +45,6 @@ fastify.register(prismaPlugin);
 fastify.get("/health", () => true);
 fastify.register(apiKeys, { prefix: "/api/keys" });
 fastify.register(chat, { prefix: "/chat" });
-fastify.register(chatContent, { prefix: "/chat/content" });
 fastify.register(chatContents, { prefix: "/chat/contents" });
 fastify.register(gpi, { prefix: "/gpi" });
 fastify.register(gpis, { prefix: "/gpis" });
